rename atlas->texture_atlas
