+++ shaders/src/texture2d.frag
@@ -1,9 +1,7 @@
 #version 450
 #extension GL_ARB_separate_shader_objects: enable
 
-// Default to 80, this gets overridden to a bigger or smaller value when the
-// shader is actually built into a graphics pipeline
-layout(constant_id = 0) const uint MAX_TEXTURES = 2;
+layout(constant_id = 0) const uint MAX_TEXTURES = 1;
 
 layout(location = 0) in vec2 vary_uv;
 layout(location = 1) in vec4 vary_rgba;
